Add doConfirmSendMessageButton.js
